For issue #16: Moved current genetic code into its own Core module
